action proto + some refactoring
