+++ token/core/fabtoken/protos/ftactions.proto
@@ -10,8 +10,48 @@ package fabtoken;
 
 option go_package = "github.com/hyperledger-labs/fabric-token-sdk/token/core/fabtoken/protos-go/actions";
 
+import "ftpp.proto";
+
 message Token {
   bytes owner = 1; // is the token owner
   string type = 2; // is the type of the token
   string quantity = 3; // is the number of units of Type carried in the token. It is encoded as a string containing a number in base 16. The string has prefix ``0x''.
+}
+
+message TokenID {
+  string id = 1;
+  uint64 index = 2;
+}
+
+message TransferActionInput {
+  TokenID token_id = 1;
+  Token input = 2;
+}
+
+message TransferActionOutput {
+  Token token = 1; // Token is the new token
+}
+
+message TransferAction {
+  uint64 version = 1;
+  repeated TransferActionInput inputs = 2; // inputs
+  repeated TransferActionOutput outputs = 3; // outputs
+  map<string, bytes> metadata = 4; // Metadata contains the transfer action's metadata
+}
+
+message IssueActionInput {
+  TokenID id = 1; // is the token id of the token to be redeemed
+  bytes token = 2; // is the actual token to be redeemed
+}
+
+message IssueActionOutput {
+  Token token = 1; // is the newly issued token
+}
+
+message IssueAction {
+  uint64 version = 1;
+  Identity issuer = 2; // is the identity of issuer
+  repeated IssueActionInput inputs = 3; // are the tokens to be redeemed by this issue action
+  repeated IssueActionOutput outputs = 4; // are the newly issued tokens
+  map<string, bytes> metadata = 5; // Metadata of the issue action
 }
